checkPriceRangeError | 최대 가격이 들어올 때 고려하여 메서드 수정

+++ src/main/java/lotto/LottoTickets.java
@@ -50,7 +50,7 @@ public class LottoTickets {
         return totalReturnPercent;
     }
 
-    private static int getPurchaseAmount() {
+    private static double getPurchaseAmount() {
         return size * ReferenceValue.Lotto.PRICE.getValue();
     }
